+++ packages/core/admin/admin/src/pages/Settings/pages/TransferTokens/CreateView.tsx
@@ -1,0 +1,16 @@
+import { CheckPagePermissions } from '@strapi/helper-plugin';
+import { useSelector } from 'react-redux';
+
+import { selectAdminPermissions } from '../../../../selectors';
+
+import { EditView } from './EditView';
+
+export const ProtectedCreateView = () => {
+  const permissions = useSelector(selectAdminPermissions);
+
+  return (
+    <CheckPagePermissions permissions={permissions.settings?.['transfer-tokens'].create}>
+      <EditView />
+    </CheckPagePermissions>
+  );
+};
