--- packages/core/admin/admin/src/pages/Settings/pages/TransferTokens/EditView/utils/tests/dataMock.js
@@ -1,14 +1,0 @@
-export const data = {
-  data: {
-    'api::address': {
-      controllers: {
-        address: ['find', 'findOne'],
-      },
-    },
-    'api::category': {
-      controllers: {
-        category: ['find', 'findOne', 'create', 'update', 'delete', 'createLocalization'],
-      },
-    },
-  },
-};
