--- packages/core/admin/admin/src/pages/Settings/pages/TransferTokens/ListView/utils/tableHeaders.js
@@ -1,48 +1,0 @@
-const tableHeaders = [
-  {
-    name: 'name',
-    key: 'name',
-    metadatas: {
-      label: {
-        id: 'Settings.tokens.ListView.headers.name',
-        defaultMessage: 'Name',
-      },
-      sortable: true,
-    },
-  },
-  {
-    name: 'description',
-    key: 'description',
-    metadatas: {
-      label: {
-        id: 'Settings.tokens.ListView.headers.description',
-        defaultMessage: 'Description',
-      },
-      sortable: false,
-    },
-  },
-  {
-    name: 'createdAt',
-    key: 'createdAt',
-    metadatas: {
-      label: {
-        id: 'Settings.tokens.ListView.headers.createdAt',
-        defaultMessage: 'Created at',
-      },
-      sortable: false,
-    },
-  },
-  {
-    name: 'lastUsedAt',
-    key: 'lastUsedAt',
-    metadatas: {
-      label: {
-        id: 'Settings.tokens.ListView.headers.lastUsedAt',
-        defaultMessage: 'Last used',
-      },
-      sortable: false,
-    },
-  },
-];
-
-export default tableHeaders;
