--- packages/core/admin/admin/src/pages/Settings/pages/TransferTokens/ProtectedCreateView/index.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-import { CheckPagePermissions } from '@strapi/helper-plugin';
-import { useSelector } from 'react-redux';
-
-import { selectAdminPermissions } from '../../../../../selectors';
-import EditView from '../EditView';
-
-const ProtectedTransferTokenCreateView = () => {
-  const permissions = useSelector(selectAdminPermissions);
-
-  return (
-    <CheckPagePermissions permissions={permissions.settings['transfer-tokens'].create}>
-      <EditView />
-    </CheckPagePermissions>
-  );
-};
-
-export default ProtectedTransferTokenCreateView;
