+++ packages/core/admin/admin/src/types/permissions.ts
@@ -5,6 +5,7 @@ type SettingsPermissions =
   | 'project-settings'
   | 'roles'
   | 'sso'
+  | 'transfer-tokens'
   | 'users'
   | 'webhooks';
