+++ packages/core/helper-plugin/src/features/Tracking.tsx
@@ -243,7 +243,7 @@ type WillModifyTokenEvent = {
   name: 'didCreateToken' | 'didEditToken';
   properties: {
     tokenType: TokenEvents['properties']['tokenType'];
-    type: 'custom' | 'full-access' | 'read-only';
+    type: 'custom' | 'full-access' | 'read-only' | Array<'push' | 'pull' | 'push-pull'>;
   };
 };
